Misc refactoring / adding comments
